feat connected to remote docker render data base

+++ SyncRazorForms/Models/Product.cs
@@ -2,9 +2,9 @@
 
 public class Product
 {
-    public int Id { get; set; } 
+    public long Id { get; set; } 
     public string? Name { get; set; }
     public string? Description { get; set; }
     public double Cost { get; set; }
-    public int Amount { get; set; }
+    public long Amount { get; set; }
 }
